Best effort upgrade for api packages

+++ api/src/graphql/utils/index.ts
@@ -1,4 +1,4 @@
-import IORedis from 'ioredis';
+import { RedisKey } from 'ioredis';
 import { redisClient } from '../../common/services';
 
 export const MATCH_SENSITIVE_CACHE_PREFIX = 'match';
@@ -9,13 +9,15 @@ export const deleteKeysByPattern = (key: string) => {
     count: 100,
   });
 
-  const keys = [] as IORedis.KeyType[];
+  const keys = [] as RedisKey[];
   stream.on('data', (resultKeys) => {
     for (let i = 0; i < resultKeys.length; i++) {
       keys.push(resultKeys[i]);
     }
   });
   stream.on('end', function () {
-    redisClient.unlink(keys);
+    if (keys.length) {
+      redisClient.unlink(keys);
+    }
   });
 };
